Compare installations

Compare installations performed into your libraries by conan.

Usage

hipercow_provision_compare(curr = 0, prev = -1, driver = NULL, root = NULL)

Arguments

curr

The previous installation to compare against. Can be a name (see hipercow_provision_list to get names), a negative number where -n indicates "n installations ago" or a positive number where n indicates "the nth installation". The default value of 0 corresponds to the current installation.

prev

The previous installation to compare against. Can be a name (see hipercow_provision_list to get names), a negative number where -n indicates "n installations ago" or a positive number where n indicates "the nth installation". The default of -1 indicates the previous installation. Must refer to an installation before curr. Use NULL or -Inf if you want to compare against the empty installation.

driver

The name of the driver to use, or you can leave blank if only one is configured (this will be typical).

root

The hipercow root

Value

An object of class conan_compare, which can be printed nicely.

Examples

cleanup <- hipercow_example_helper()
#> ℹ This example uses a special helper
hipercow_provision("pkgdepends", refs = "knitr")
#> /`-'\  _______  ___  ___ ____
#> \,T./ / __/ _ \/ _ \/ _ `/ _ \
#>   |   \__/\___/_//_/\_,_/_//_/
#>   |   ---- THE  LIBRARIAN ----
#> 
#> Bootstrapping from: /home/runner/work/_temp/Library
#> Installing into library: hipercow/lib
#> Using method pkgdepends
#> Running in path: /tmp/RtmpIKAu8X/file18063a2a731f
#> Library paths:
#>   - /tmp/RtmpIKAu8X/file18063a2a731f/hipercow/lib
#>   - /opt/R/4.3.2/lib/R/site-library
#>   - /opt/R/4.3.2/lib/R/library
#> id: 20240128222700
#> Logs from pkgdepends follow:
#> 
#> -------------------------------------------------------------------------------
#> 
#> 
#> ── repos 
#> • https://cloud.r-project.org
#> 
#> ── refs 
#> • knitr
#> 
#> ✔ Updated metadata database: 4.19 MB in 9 files.
#> 
#> ℹ Updating metadata database
#> ✔ Updating metadata database ... done
#> 
#> + evaluate   0.23  [bld][dl]
#> + highr      0.10  [bld][dl]
#> + knitr      1.45  [bld][dl]
#> + xfun       0.41  [bld][cmp][dl]
#> + yaml       2.3.8 [bld][cmp][dl]
#> ℹ Getting 5 pkgs with unknown sizes
#> ✔ Got evaluate 0.23 (source) (28.35 kB)
#> ✔ Got highr 0.10 (source) (15.08 kB)
#> ✔ Got yaml 2.3.8 (source) (94.76 kB)
#> ✔ Got xfun 0.41 (source) (135.72 kB)
#> ✔ Got knitr 1.45 (source) (898.76 kB)
#> ℹ Building evaluate 0.23
#> ℹ Building xfun 0.41
#> ℹ Building yaml 2.3.8
#> ✔ Built evaluate 0.23 (1.7s)
#> ✔ Installed evaluate 0.23  (31ms)
#> ✔ Built xfun 0.41 (3.1s)
#> ✔ Installed xfun 0.41  (1s)
#> ℹ Building highr 0.10
#> ✔ Built yaml 2.3.8 (5.1s)
#> ✔ Installed yaml 2.3.8  (1s)
#> ✔ Built highr 0.10 (2s)
#> ✔ Installed highr 0.10  (1s)
#> ℹ Building knitr 1.45
#> ✔ Built knitr 1.45 (4.5s)
#> ✔ Installed knitr 1.45  (1s)
#> ✔ Summary:   5 new  in 20.6s
#> 
#> -------------------------------------------------------------------------------
#> Writing library description to 'hipercow/lib/.conan/20240128222700'
#> Done!
hipercow_provision("pkgdepends", refs = "data.table")
#> /`-'\  _______  ___  ___ ____
#> \,T./ / __/ _ \/ _ \/ _ `/ _ \
#>   |   \__/\___/_//_/\_,_/_//_/
#>   |   ---- THE  LIBRARIAN ----
#> 
#> Bootstrapping from: /home/runner/work/_temp/Library
#> Installing into library: hipercow/lib
#> Using method pkgdepends
#> Running in path: /tmp/RtmpIKAu8X/file18063a2a731f
#> Library paths:
#>   - /tmp/RtmpIKAu8X/file18063a2a731f/hipercow/lib
#>   - /opt/R/4.3.2/lib/R/site-library
#>   - /opt/R/4.3.2/lib/R/library
#> id: 20240128222721
#> Logs from pkgdepends follow:
#> 
#> -------------------------------------------------------------------------------
#> 
#> 
#> ── repos 
#> • https://cloud.r-project.org
#> 
#> ── refs 
#> • data.table
#> ℹ Loading metadata database
#> ✔ Loading metadata database ... done
#> 
#> + data.table   1.14.10 [bld][cmp]
#> ℹ No downloads are needed, 1 pkg is cached
#> ✔ Got data.table 1.14.10 (source) (5.31 MB)
#> ℹ Building data.table 1.14.10
#> ✔ Built data.table 1.14.10 (15.9s)
#> ✔ Installed data.table 1.14.10  (1s)
#> ✔ Summary:   1 new  in 17s
#> 
#> -------------------------------------------------------------------------------
#> Writing library description to 'hipercow/lib/.conan/20240128222721'
#> Done!
hipercow_provision_compare()
#> ── Comparing conan installations ───────────────────────────────────────────────
#> • 20240128222700 1st; previous installation (less than a minute ago)
#> • 20240128222721 2nd; current installation (moments ago)
#> 
#> ── 5 unchanged packages ──
#> 
#> ℹ To show unchanged packages, print with 'show_unchanged = TRUE'
#> 
#> ── 1 added package ──
#> 
#> • data.table (1.14.10) CRAN

cleanup()
#> ℹ Cleaning up example