List installations

List previous successful installations of this hipercow root.

Usage

hipercow_provision_list(driver = NULL, root = NULL)

hipercow_provision_check(
  method = NULL,
  ...,
  driver = NULL,
  environment = "default",
  root = NULL
)

Arguments

driver

The name of the driver to use, or you can leave blank if only one is configured (this will be typical).

root

The hipercow root

method

The provisioning method to use, defaulting to NULL, which indicates we should try and detect the best provisioning mechanism for you; this should typically work well unless you are manually adding packages into your library (see Details). If given, must be one of auto, pkgdepends, script or renv; each of these are described in the Details and in vignette("packages").

...

Arguments passed through to conan. See Details.

environment

The name of the environment to provision (see hipercow_environment_create for details).

Value

A data.frame with columns:

name: the name of the installation. This might be useful with conan_compare

time: the time the installation was started

hash: the installation hash

method: the method used for the installation

args: the arguments to the installation (as a list column)

current: if using hipercow_provision_check, does this installation match the arguments provided?

This object also has class conan_list so that it prints nicely, but you can drop this with as.data.frame.

Examples

cleanup <- hipercow_example_helper()
#> ℹ This example uses a special helper
writeLines("data.table", "pkgdepends.txt")

# Before any installation has happened:
hipercow_provision_list()
#> ! No conan installations recorded
hipercow_provision_check()
#> ✔ Selected provisioning method 'pkgdepends'
#> ! No conan installations recorded

# After installation:
hipercow_provision()
#> ✔ Selected provisioning method 'pkgdepends'
#> /`-'\  _______  ___  ___ ____
#> \,T./ / __/ _ \/ _ \/ _ `/ _ \
#>   |   \__/\___/_//_/\_,_/_//_/
#>   |   ---- THE  LIBRARIAN ----
#> 
#> Bootstrapping from: /home/runner/work/_temp/Library
#> Installing into library: hipercow/lib
#> Using method pkgdepends
#> Running in path: /tmp/Rtmp2u5dKy/file17c328203efe
#> Library paths:
#>   - /tmp/Rtmp2u5dKy/file17c328203efe/hipercow/lib
#>   - /opt/R/4.3.2/lib/R/site-library
#>   - /opt/R/4.3.2/lib/R/library
#> id: 20240201163317
#> Logs from pkgdepends follow:
#> 
#> -------------------------------------------------------------------------------
#> 
#> 
#> ── repos 
#> • https://cloud.r-project.org
#> 
#> ── refs 
#> • data.table
#> 
#> ✔ Updated metadata database: 4.20 MB in 9 files.
#> 
#> ℹ Updating metadata database
#> ✔ Updating metadata database ... done
#> 
#> ! Failed to update system requirement mappings, will use cached mappings.
#> + data.table   1.15.0 [bld][cmp][dl]
#> ℹ Getting 1 pkg with unknown size
#> ✔ Got data.table 1.15.0 (source) (5.39 MB)
#> ℹ Building data.table 1.15.0
#> ✔ Built data.table 1.15.0 (17.1s)
#> ✔ Installed data.table 1.15.0  (1s)
#> ✔ Summary:   1 new  in 18.2s
#> 
#> -------------------------------------------------------------------------------
#> Writing library description to 'hipercow/lib/.conan/20240201163317'
#> Done!
hipercow_provision_list()
#> ℹ 1 conan installation recorded
#> • 1: 20240201163317 (moments ago) [0]
hipercow_provision_check()
#> ✔ Selected provisioning method 'pkgdepends'
#> ℹ 1 conan installation recorded
#> • 1: 20240201163317 (moments ago) [0] (*)
#> ℹ The entry marked with '*' matches the provided installation hash

# After a different installation:
hipercow_provision("pkgdepends", refs = "knitr")
#> /`-'\  _______  ___  ___ ____
#> \,T./ / __/ _ \/ _ \/ _ `/ _ \
#>   |   \__/\___/_//_/\_,_/_//_/
#>   |   ---- THE  LIBRARIAN ----
#> 
#> Bootstrapping from: /home/runner/work/_temp/Library
#> Installing into library: hipercow/lib
#> Using method pkgdepends
#> Running in path: /tmp/Rtmp2u5dKy/file17c328203efe
#> Library paths:
#>   - /tmp/Rtmp2u5dKy/file17c328203efe/hipercow/lib
#>   - /opt/R/4.3.2/lib/R/site-library
#>   - /opt/R/4.3.2/lib/R/library
#> id: 20240201163344
#> Logs from pkgdepends follow:
#> 
#> -------------------------------------------------------------------------------
#> 
#> 
#> ── repos 
#> • https://cloud.r-project.org
#> 
#> ── refs 
#> • knitr
#> ℹ Loading metadata database
#> ✔ Loading metadata database ... done
#> 
#> + evaluate   0.23  [bld]
#> + highr      0.10  [bld]
#> + knitr      1.45  [bld]
#> + xfun       0.41  [bld][cmp]
#> + yaml       2.3.8 [bld][cmp]
#> ℹ No downloads are needed, 5 pkgs are cached
#> ✔ Got evaluate 0.23 (source) (28.35 kB)
#> ✔ Got highr 0.10 (source) (15.08 kB)
#> ✔ Got yaml 2.3.8 (source) (94.76 kB)
#> ✔ Got xfun 0.41 (source) (135.72 kB)
#> ✔ Got knitr 1.45 (source) (898.76 kB)
#> ℹ Building evaluate 0.23
#> ℹ Building xfun 0.41
#> ℹ Building yaml 2.3.8
#> ✔ Built evaluate 0.23 (1.8s)
#> ✔ Installed evaluate 0.23  (65ms)
#> ✔ Built xfun 0.41 (3.2s)
#> ✔ Installed xfun 0.41  (1s)
#> ℹ Building highr 0.10
#> ✔ Built yaml 2.3.8 (5.1s)
#> ✔ Installed yaml 2.3.8  (30ms)
#> ✔ Built highr 0.10 (1.1s)
#> ✔ Installed highr 0.10  (1s)
#> ℹ Building knitr 1.45
#> ✔ Built knitr 1.45 (4.5s)
#> ✔ Installed knitr 1.45  (1s)
#> ✔ Summary:   5 new  in 18.9s
#> 
#> -------------------------------------------------------------------------------
#> Writing library description to 'hipercow/lib/.conan/20240201163344'
#> Done!
hipercow_provision_list()
#> ℹ 2 conan installations recorded
#> • 1: 20240201163317 (less than a minute ago) [-1]
#> • 2: 20240201163344 (moments ago) [0]
hipercow_provision_check()
#> ✔ Selected provisioning method 'pkgdepends'
#> ℹ 2 conan installations recorded
#> • 1: 20240201163317 (less than a minute ago) [-1] (*)
#> • 2: 20240201163344 (moments ago) [0]
#> ℹ The entry marked with '*' matches the provided installation hash

cleanup()
#> ℹ Cleaning up example