Validate a hipercow_resources list for a driver.

Query a driver to find information about the cluster, and then validate a hipercow_resources list against that driver to see if the resources requested could be satisfied.

Usage

hipercow_resources_validate(resources, driver = NULL, root = NULL)

Arguments

resources

A hipercow_resources list returned by hipercow_resources, or NULL

driver

The name of the driver to use, or you can leave blank if only one is configured (this will be typical).

root

The hipercow root

Value

TRUE if the resources are compatible with this driver.

Examples

cleanup <- hipercow_example_helper()
#> ℹ This example uses a special helper
hipercow_resources_validate(hipercow_resources(cores = 1))
#> $cores
#> $cores$original
#> [1] 1
#> 
#> $cores$computed
#> [1] 1
#> 
#> 
#> $exclusive
#> $exclusive$original
#> [1] FALSE
#> 
#> $exclusive$computed
#> [1] FALSE
#> 
#> 
#> $max_runtime
#> $max_runtime$original
#> NULL
#> 
#> $max_runtime$computed
#> NULL
#> 
#> 
#> $hold_until
#> $hold_until$original
#> NULL
#> 
#> $hold_until$computed
#> NULL
#> 
#> 
#> $memory_per_node
#> $memory_per_node$original
#> NULL
#> 
#> $memory_per_node$computed
#> NULL
#> 
#> 
#> $memory_per_process
#> $memory_per_process$original
#> NULL
#> 
#> $memory_per_process$computed
#> NULL
#> 
#> 
#> $requested_nodes
#> $requested_nodes$original
#> NULL
#> 
#> $requested_nodes$computed
#> NULL
#> 
#> 
#> $priority
#> $priority$original
#> NULL
#> 
#> $priority$computed
#> NULL
#> 
#> 
#> $queue
#> $queue$original
#> NULL
#> 
#> $queue$computed
#> [1] "default"
#> 
#> 
#> attr(,"class")
#> [1] "hipercow_resources"
#> attr(,"validated")
#> [1] "example"

# This example does not allow more than one core
tryCatch(
  hipercow_resources_validate(hipercow_resources(cores = 32)),
  error = identity)
#> <error/rlang_error>
#> Error in `validate_cluster_cores()`:
#> ! 32 is too many cores for this cluster.
#> ℹ The largest node has 1 cores.
#> ---
#> Backtrace:
#>      ▆
#>   1. └─pkgdown::build_site_github_pages(new_process = FALSE, install = FALSE)
#>   2.   └─pkgdown::build_site(...)
#>   3.     └─pkgdown:::build_site_local(...)
#>   4.       └─pkgdown::build_reference(...)
#>   5.         └─purrr::map(...)
#>   6.           └─purrr:::map_("list", .x, .f, ..., .progress = .progress)
#>   7.             ├─purrr:::with_indexed_errors(...)
#>   8.             │ └─base::withCallingHandlers(...)
#>   9.             ├─purrr:::call_with_cleanup(...)
#>  10.             └─pkgdown (local) .f(.x[[i]], ...)
#>  11.               ├─base::withCallingHandlers(...)
#>  12.               └─pkgdown:::data_reference_topic(...)
#>  13.                 └─pkgdown:::run_examples(...)
#>  14.                   └─pkgdown:::highlight_examples(code, topic, env = env)
#>  15.                     └─downlit::evaluate_and_highlight(...)
#>  16.                       └─evaluate::evaluate(code, child_env(env), new_device = TRUE, output_handler = output_handler)
#>  17.                         └─evaluate:::evaluate_call(...)
#>  18.                           ├─evaluate (local) timing_fn(...)
#>  19.                           ├─evaluate (local) handle(...)
#>  20.                           │ └─base::try(f, silent = TRUE)
#>  21.                           │   └─base::tryCatch(...)
#>  22.                           │     └─base (local) tryCatchList(expr, classes, parentenv, handlers)
#>  23.                           │       └─base (local) tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>  24.                           │         └─base (local) doTryCatch(return(expr), name, parentenv, handler)
#>  25.                           ├─base::withCallingHandlers(...)
#>  26.                           ├─base::withVisible(...)
#>  27.                           └─evaluate:::eval_with_user_handlers(expr, envir, enclos, user_handlers)
#>  28.                             └─base::eval(expr, envir, enclos)
#>  29.                               └─base::eval(expr, envir, enclos)
#>  30.                                 ├─base::tryCatch(...)
#>  31.                                 │ └─base (local) tryCatchList(expr, classes, parentenv, handlers)
#>  32.                                 │   └─base (local) tryCatchOne(expr, names, parentenv, handlers[[1L]])
#>  33.                                 │     └─base (local) doTryCatch(return(expr), name, parentenv, handler)
#>  34.                                 └─hipercow::hipercow_resources_validate(hipercow_resources(cores = 32))
#>  35.                                   └─hipercow:::resources_validate(resources, driver, root)
#>  36.                                     └─hipercow:::validate_cluster_cores(resources$cores$computed, cluster_resources$max_cores)

cleanup()
#> ℹ Cleaning up example