Create bulk tasks via with expressions

Create a bulk set of tasks. This is an experimental interface and does not have an analogue within didehpc. Variables in data take precidence over variables in the environment in which expr was created. There is no "pronoun" support yet (see rlang docs). Use !! to pull a variable from the environment if you need to, but be careful not to inject something really large (e.g., any vector really) or you'll end up with a revolting expression and poor backtraces. We will likely change some of these semantics later, be careful.

task_create_bulk_expr(
  expr,
  data,
  environment = "default",
  submit = NULL,
  root = NULL
)

Arguments

expr

An expression, as for task_create_expr

data

Data that you wish to inject rowwise into the expression

environment

Name of the hipercow environment to evaluate the task within.

submit

Control over task submission. This will expand over time once we support specifying resources. The most simple interface is to use TRUE here to automatically submit a task, using your default configuration, or FALSE to prevent submission. The default NULL will submit a task if a driver is configured.

root

A hipercow root, or path to it. If NULL we search up your directory tree.

Value

A vector of ids, with the same length as data has rows.