Create explicit task

Create an explicit task. Explicit tasks are the simplest sort of task in hipercow and do nothing magic. They accept an R expression (from quote or friends) and possibly a set of variables to export from the global environment. This can then be run on a cluster by loading your variables and running your expression. If your expression depends on packages being attached then you should pass a vector of package names too. This function may disappear, and is used by us to think about the package, it's not designed to really be used.

task_create_explicit(
  expr,
  export = NULL,
  envir = .GlobalEnv,
  environment = "default",
  submit = NULL,
  root = NULL
)

Arguments

expr

Unevaluated expression object, e.g., from quote

export

Optional character vector of names of objects to export into the evaluating environment

envir

Local R environment in which to find variables for export

environment

Name of the hipercow environment to evaluate the task within.

submit

Control over task submission. This will expand over time once we support specifying resources. The most simple interface is to use TRUE here to automatically submit a task, using your default configuration, or FALSE to prevent submission. The default NULL will submit a task if a driver is configured.

root

A hipercow root, or path to it. If NULL we search up your directory tree.

Value

A task id, a string of hex characters. Use this to interact with the task.