Run a task

Run a task that has been created by a task_create_* function, e.g., task_create_explicit(), task_create_expr(). Generally users should not run this function directly.

Usage

task_eval(id, envir = .GlobalEnv, verbose = FALSE, root = NULL)

Arguments

id

The task identifier

envir

An environment in which to evaluate the expression. For non-testing purposes, generally ignore this, the global environment will be likely the expected environment.

verbose

Logical, indicating if we should print information about what we do as we do it.

root

A hipercow root, or path to it. If NULL we search up your directory tree.

Value

Logical indicating success (TRUE) or failure (FALSE)

Examples

cleanup <- hipercow_example_helper(runner = FALSE)
#> ℹ This example uses a special helper
id <- task_create_expr(runif(1), driver = FALSE)
# Status is only 'created', not 'submitted', as we did not submit
# task.  This task can never run.
task_status(id)
#> [1] "created"

# Explicitly evaluate the task:
task_eval(id, verbose = TRUE)
#> 
#> ── hipercow 1.0.12 running at '/tmp/RtmpsGF8Lo/file196316355766' ───────────────
#> ℹ library paths:
#> • /home/runner/work/_temp/Library
#> • /opt/R/4.3.3/lib/R/site-library
#> • /opt/R/4.3.3/lib/R/library
#> ℹ id: 83f620631074f4d12c8ec6e7f9d99ee2
#> ℹ starting at: 2024-04-23 08:22:42.751971
#> ℹ Task type: expression
#>   • Expression: runif(1)
#>   • Locals: (none)
#>   • Environment: default
#> ───────────────────────────────────────────────────────────────── task logs ↓ ──
#> 
#> ───────────────────────────────────────────────────────────────── task logs ↑ ──
#> ✔ status: success
#> ℹ finishing at: 2024-04-23 08:22:42.751971 (elapsed: 0.1567 secs)
task_result(id)
#> [1] 0.5144025

cleanup()
#> ℹ Cleaning up example