Get task log

Get the task log, if the task has produced one. Tasks run by the windows driver will generally produce a log. A log might be quite long, and you might want to print it to screen in its entirety (task_log_show), or return it as character vector (task_log_value).

Usage

task_log_show(id, follow = TRUE, outer = FALSE, root = NULL)

task_log_value(id, follow = TRUE, outer = FALSE, root = NULL)

task_log_watch(
  id,
  follow = TRUE,
  poll = 1,
  skip = 0,
  timeout = NULL,
  progress = NULL,
  root = NULL
)

Arguments

id

The task identifier

follow

Logical, indicating if we should follow any retried tasks.

outer

Logical, indicating if we should request the "outer" logs; these are logs from the underlying HPC software before it hands off to hipercow.

root

A hipercow root, or path to it. If NULL we search up your directory tree.

poll

Time, in seconds, used to throttle calls to the status function. The default is 1 second

skip

Optional integer indicating how to handle log content that exists at the point where we start watching. The default (0) shows all log contents. A positive integer skips that many lines, while a negative integer shows only that many lines (so -5 shows the first five lines in the log). You can pass Inf to discard all previous logs, but stream all new ones.

timeout

The time to wait for the task to complete. The default is to wait forever.

progress

Logical value, indicating if a progress spinner should be used. The default NULL uses the option hipercow.progress, and if unset displays a progress bar in an interactive session.

Value

Depending on the function:

task_log_show returns the log value contents invisibly, but primarily displays the log contents on the console as a side effect

task_log_value returns a character of log contents

task_log_watch returns the status converted to logical (as for task_wait)

Details

The function task_log_watch has similar semantics to task_wait but does not error on timeout, and always displays a log.

Examples

cleanup <- hipercow_example_helper(with_logging = TRUE)
#> ℹ This example uses a special helper

# Tasks that don't produce any output (print, cat, warning, etc)
# will only contain logging information from hipercow itself
id <- task_create_expr(runif(1))
#> ✔ Submitted task '0878e043a638e3dc3fe6053c4d5cbc09' using 'example'
task_wait(id)
#> [1] TRUE
task_log_show(id)
#> 
#> ── hipercow 0.3.3 running at '/tmp/RtmpqBs51w/file17ec9945873' ─────────────────
#> ℹ library paths:
#> • /home/runner/work/_temp/Library
#> • /opt/R/4.3.2/lib/R/site-library
#> • /opt/R/4.3.2/lib/R/library
#> ℹ id: 0878e043a638e3dc3fe6053c4d5cbc09
#> ℹ starting at: 2024-02-07 09:14:00.662361
#> ℹ Task type: expression
#>   • Expression: runif(1)
#>   • Locals: (none)
#>   • Environment: default
#> ───────────────────────────────────────────────────────────────── task logs ↓ ──
#> 
#> ───────────────────────────────────────────────────────────────── task logs ↑ ──
#> ✔ status: success
#> ℹ finishing at: 2024-02-07 09:14:00.662361 (elapsed: 0.2139 secs)

# If your task creates output then it will appear within the
# horizontal rules:
id <- task_create_expr({
  message("Starting analysis")
  x <- mean(runif(100))
  message("all done!")
  x
})
#> ✔ Submitted task '4977f29ffb66938ac96b983429ec2d26' using 'example'
task_wait(id)
#> [1] TRUE
task_log_show(id)
#> 
#> ── hipercow 0.3.3 running at '/tmp/RtmpqBs51w/file17ec9945873' ─────────────────
#> ℹ library paths:
#> • /home/runner/work/_temp/Library
#> • /opt/R/4.3.2/lib/R/site-library
#> • /opt/R/4.3.2/lib/R/library
#> ℹ id: 4977f29ffb66938ac96b983429ec2d26
#> ℹ starting at: 2024-02-07 09:14:01.486333
#> ℹ Task type: expression
#>   • Expression: { [...]
#>   • Locals: (none)
#>   • Environment: default
#> ───────────────────────────────────────────────────────────────── task logs ↓ ──
#> Starting analysis
#> all done!
#> 
#> ───────────────────────────────────────────────────────────────── task logs ↑ ──
#> ✔ status: success
#> ℹ finishing at: 2024-02-07 09:14:01.486333 (elapsed: 0.2253 secs)

# Use "task_log_value" to get the log value as a character vector
task_log_value(id)
#>  [1] ""                                                                                
#>  [2] "── hipercow 0.3.3 running at '/tmp/RtmpqBs51w/file17ec9945873' ─────────────────"
#>  [3] "ℹ library paths:"                                                                
#>  [4] "• /home/runner/work/_temp/Library"                                               
#>  [5] "• /opt/R/4.3.2/lib/R/site-library"                                               
#>  [6] "• /opt/R/4.3.2/lib/R/library"                                                    
#>  [7] "ℹ id: 4977f29ffb66938ac96b983429ec2d26"                                          
#>  [8] "ℹ starting at: 2024-02-07 09:14:01.486333"                                       
#>  [9] "ℹ Task type: expression"                                                         
#> [10] "  • Expression: { [...]"                                                         
#> [11] "  • Locals: (none)"                                                              
#> [12] "  • Environment: default"                                                        
#> [13] "───────────────────────────────────────────────────────────────── task logs ↓ ──"
#> [14] "Starting analysis"                                                               
#> [15] "all done!"                                                                       
#> [16] ""                                                                                
#> [17] "───────────────────────────────────────────────────────────────── task logs ↑ ──"
#> [18] "✔ status: success"                                                               
#> [19] "ℹ finishing at: 2024-02-07 09:14:01.486333 (elapsed: 0.2253 secs)"               

# Depending on the driver you are using, there may be useful
# information in the "outer" log; the logs produced by the
# submission system before hipercow takes over:
task_log_show(id, outer = TRUE)
#> Running task 4977f29ffb66938ac96b983429ec2d26
#> Finished task 4977f29ffb66938ac96b983429ec2d26

cleanup()
#> ℹ Cleaning up example