Submit a task

Submit a task to a queue. This is a lower-level function that you will not often need to call. Typically a task will be submitted automatically to your driver on creation (e.g., with task_create_expr()), unless you specified driver = FALSE or you had not yet configured a driver.

Usage

task_submit(id, ..., resources = NULL, driver = NULL, root = NULL)

Arguments

id

A vector of task ids

...

Disallowed additional arguments, don't use.

resources

A list generated by hipercow_resources giving the cluster resource requirements to run your task.

driver

The name of the driver to use, or you can leave blank if only one is configured (this will be typical).

root

The hipercow root